python_services/pyproject.toml
[project]
name = "python-services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "django-environ>=0.12.0",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "djangorestframework>=3.16.1",
    "scikit-learn>=1.7.2",
    "numpy>=2.3.3",
    "torch>=2.8.0",
    "requests>=2.31.0",
    "python3-discogs-client>=2.3.0",
    "anthropic>=0.18.0",
    "gunicorn>=21.2.0",
    "django-cors-headers>=4.0.0",
    "pytest-django>=4.11.1",
]
[tool.setuptools]
packages = ["bandit", "config"]
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings_test"

[dependency-groups]
dev = [
    "pytest<9",
]
